+++ fastmvc_db_models/factories/__init__.py
@@ -1,0 +1,11 @@
+"""
+factory_boy examples for tests (install ``pip install 'fastmvc-db-models[dev]'``).
+
+::
+
+    from fastmvc_db_models.factories import PlanFactory
+"""
+
+from fastmvc_db_models.factories.plan_factory import PlanFactory
+
+__all__ = ["PlanFactory"]