+++ tests/__init__.py
@@ -1,1 +1,1 @@
-"""Tests."""
+"""Tests for fastmvc_db_models."""